More information sent along with a submission.
There are now three things sent in the header section of a submission; A magic
character sequence shared by the client and server to indicate the end of a
message, the paricipant's name, and the name of the challenge.

Also fixed a bug where the client may misinterpret the magic character sequence
due to it being split on 1024 byte boundaries.