Fixed the problem with Manager connection reading from its socket twice.
Here's how it was before:

  Participants are all ready
  Boss blocks on manager connection while manager connection waits to hear a yes
  or no from manager to start the challenge
  Boss then tells every participant to start the challenge.

The problem was that the action of manager connection waiting to hear from
manager on its socket was that the manager connection thread was already
listening on the socket, so it intercepted the message manager connection wanted
to see.

Here's the fix:
  Participants are all ready
  Boss tells manager connection to tell the manager the challenge is ready

That's it. Eventually the manager connection thread will get the response from
the manager and then tell the boss to tell the participants to either start the
challenge or cancel the challenge.